Updated ham/xnec2c 3.4 to 3.5 ----------------------------- (from doc/xnec2c.html) Version 3.5:
I modified the NEC2 Editor code so that when a Geometry or
Command editor window is opened, activating Apply or OK will
save data in the editor window to the treeview, even if the
default data is not edited.

I modified the NEC2 Editor code so that if a treeview row is
removed while the relevant editor window is open, then
activating Apply or OK will not attempt to save data to the
treeview, since this will cause an illegal memory access and
will crash xnec2c.

Fixed a bug in the Helix editor which caused incorrect
calculation of the segment length as a percentage of
wavelength or of the number of segments/turn that correspond
to a given segment length as a percentage of wavelength.