Changes 3.15.0: Added support for row values. Allow deterministic SQL functions in the WHERE clause of a partial index. Added the "modeof=filename" URI parameter on the unix VFS Added support for SQLITE_DBCONFIG_MAINDBNAME. Added the ability to VACUUM an ATTACH-ed database. Enhancements to the command-line shell: Add the ".testcase" and ".check" dot-commands. Added the --new option to the ".open" dot-command, causing any prior content in the database to be purged prior to opening. Enhance the fts5vocab virtual table to handle "ORDER BY term" efficiently. Miscellaneous micro-optimizations reduce CPU usage by more than 7% on common workloads. Most optimization in this release has been on the front-end (sqlite3_prepare_v2()).
Bug Fixes:

The multiply operator now correctly detects 64-bit integer overflow and promotes to floating point in all corner-cases.
Correct handling of columns with redundant unique indexes when those columns are used on the LHS of an IN operator.
Skip NULL entries on range queries in indexes on expressions.
Ensure that the AUTOINCREMENT counters in the sqlite_sequence table are initialized doing "Xfer Optimization" on "INSERT ... SELECT" statements.
Make sure the ORDER BY LIMIT optimization (from check-in 559733b09e) works with IN operators on INTEGER PRIMARY KEYs.